+++ src/components/PostHeader/PostHeader.jsx
@@ -12,19 +12,33 @@ const PostHeader = props => {
   };
 
   const postDelete = () => {
-    fetch('/data/data.json', {
-      method: 'DELETE',
-      headers: {
-        'Content-Type': 'application/json;charset=utf-8',
+    if (window.confirm('정말 삭제하시겠습니까?')) {
+      fetch('/data/data.json', {
+        method: 'DELETE',
         headers: {
-          Authorization: localStorage.getItem('access_token'),
+          'Content-Type': 'application/json;charset=utf-8',
+          headers: {
+            Authorization: localStorage.getItem('access_token'),
+          },
         },
-      },
-      // TODO api 업데이트 후 적용
-      body: JSON.stringify({}),
-    }).then(() => {
-      props.fetchPost();
-    });
+        // TODO api 업데이트 후 적용
+        body: JSON.stringify({}),
+      })
+        .then(res => {
+          if (res.ok) {
+            alert('삭제되었습니다!');
+            props.fetchPost();
+            return;
+          }
+          if (!res.ok) {
+            throw new Error('Error');
+          }
+        })
+
+        .catch(() => {
+          alert('삭제에 실..패..했습니다....');
+        });
+    }
   };
 
   return (
